plugins need to instantiate optparse.Option
The Check.add_option method currently requires its argument to be of
optparse.Option type. This adds a requirement for all checks to use this
library to import optparse and create an Option object directly. This in
fact defeats the objective of abstracting option parsing from the
checks.

Remove this dependancy by passing on all of add_option's arguments to
OptionParser.add_option.

Correct the example check_ping check to reflect this change.

In check_ping, change the error message on import error to make it look
more like a Nagios status, make it output to stdout instead of stderr so
that Nagios catches it as the plugin's output and change the exit code
to 3 so that it corresponds to an UNKNOWN status in the Nagios plugin
specification.